requirements.txt
-i https://pypi.org/simple
alembic==1.14.1; python_version >= '3.8'
aniso8601==10.0.0
appnope==0.1.4; python_version >= '3.6'
asttokens==3.0.0; python_version >= '3.8'
backcall==0.2.0
bcrypt==4.3.0; python_version >= '3.8'
click==8.1.8; python_version >= '3.7'
decorator==5.2.1; python_version >= '3.8'
exceptiongroup==1.2.2; python_version >= '3.7'
executing==2.2.0; python_version >= '3.8'
faker==35.2.2; python_version >= '3.8'
flask==2.2.5; python_version >= '3.7'
flask-bcrypt==1.0.1
flask-migrate==4.1.0; python_version >= '3.6'
flask-restful==0.3.10
flask-sqlalchemy==3.0.3; python_version >= '3.7'
importlib-metadata==8.5.0; python_version >= '3.8'
importlib-resources==6.4.5; python_version >= '3.8'
iniconfig==2.1.0; python_version >= '3.8'
ipdb==0.13.9; python_version >= '2.7'
ipython==8.12.3; python_version >= '3.8'
itsdangerous==2.2.0; python_version >= '3.8'
jedi==0.19.2; python_version >= '3.6'
jinja2==3.1.6; python_version >= '3.7'
mako==1.3.9; python_version >= '3.8'
markupsafe==2.1.5; python_version >= '3.7'
matplotlib-inline==0.1.7; python_version >= '3.8'
packaging==24.2; python_version >= '3.8'
parso==0.8.4; python_version >= '3.6'
pexpect==4.9.0; sys_platform != 'win32'
pickleshare==0.7.5
pluggy==1.5.0; python_version >= '3.8'
prompt-toolkit==3.0.50; python_full_version >= '3.8.0'
ptyprocess==0.7.0
pure-eval==0.2.3
pygments==2.19.1; python_version >= '3.8'
pytest==8.3.5; python_version >= '3.8'
pytest-flask==1.3.0; python_version >= '3.7'
python-dateutil==2.9.0.post0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2'
python-dotenv==1.0.1; python_version >= '3.8'
pytz==2025.1
setuptools==75.3.2; python_version >= '3.8'
six==1.17.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2'
sqlalchemy==2.0.39; python_version >= '3.7'
stack-data==0.6.3
toml==0.10.2; python_version >= '2.6' and python_version not in '3.0, 3.1, 3.2'
tomli==2.2.1; python_version >= '3.8'
traitlets==5.14.3; python_version >= '3.8'
typing-extensions==4.12.2; python_version >= '3.8'
wcwidth==0.2.13
werkzeug==2.2.2; python_version >= '3.7'
zipp==3.20.2; python_version >= '3.8'
